+++ src/main/ts/view/value/a-palette.ts
