+++ src/main/ts/view/value/checkbox.ts
