+++ src/main/ts/view/value/color-component-texts.ts
