+++ src/main/ts/view/value/color-picker.ts
