+++ src/main/ts/view/value/color-swatch-text.ts
