+++ src/main/ts/view/value/color-swatch.ts
