+++ src/main/ts/view/value/h-palette.ts
