+++ src/main/ts/view/value/list.ts
