+++ src/main/ts/view/value/point-2d-pad.ts
