+++ src/main/ts/view/value/point-2d-text.ts
