+++ src/main/ts/view/value/slider-text.ts
