+++ src/main/ts/view/value/slider.ts
