+++ src/main/ts/view/value/sv-palette.ts
