+++ src/main/ts/view/value/text.ts
