+++ src/main/ts/view/value/value.ts
